DataCollector/pyproject.toml
[tool.poetry]
name = "datacollector"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
paho-mqtt = "^2.1.0"
requests = "^2.32.3"
pydantic-settings = "^2.7.1"
pydantic = "^2.10.5"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
poethepoet = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.poe.tasks]
start = "python -B datacollector/main.py"